+++ src/main/java/TaskTracker/database/map/TaskMapper.java
@@ -12,12 +12,12 @@ public class TaskMapper implements RowMapper<Task> {
 
     public Task mapRow(ResultSet resultSet, int rowNumber) throws SQLException {
         Task task = new Task();
-        task.setTaskID(resultSet.getInt("taskid"));
+        task.setTaskID(resultSet.getLong("taskid"));
         task.setCreatorLogin(resultSet.getString("creatorLogin"));
-        task.setCreatorGroupID(resultSet.getInt("creatorgroupid"));
+        task.setCreatorGroupID(resultSet.getLong("creatorgroupid"));
         task.setTaskName(resultSet.getString("taskname"));
         task.setTaskDescription(resultSet.getString("taskdescription"));
-        task.setTaskPriority(resultSet.getInt("taskpriority"));
+        task.setTaskPriority(resultSet.getLong("taskpriority"));
         task.setFinished(resultSet.getBoolean("taskstatus"));
         task.setTaskExpiryDate(resultSet.getTimestamp("taskexpirydate"));
         return task;
